Created dataclass ShapResult to store shap explanations there and encapsulate the logic of uploading SHAP charts to the WandB.
